bimba3d_backend/requirements.windows.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
email-validator
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4
sqlalchemy
psycopg[binary]
alembic
numpy
Pillow
imageio
tqdm
opencv-python
pycolmap
piexif
scikit-learn
scipy
matplotlib
pyyaml
torchmetrics
tensorboard
tyro
rich
